+++ packages/ui/Auth/Layout/index.tsx
@@ -1,0 +1,1 @@
+export {default as Layout} from './Layout';
