+++ packages/ui/Auth/LogIn/Form.tsx
@@ -1,5 +1,5 @@
 import { useForm, SubmitHandler } from 'react-hook-form'
-import { literal, object, string, TypeOf } from 'zod'
+import { object, string, TypeOf } from 'zod'
 import { zodResolver } from '@hookform/resolvers/zod'
 import { Fragment, useEffect, useState } from 'react'
 import { LoadingButton } from '@mui/lab'
@@ -11,30 +11,18 @@ import Typography from '@mui/material/Typography'
 import TextField from '@mui/material/TextField'
 import Button from '@mui/material/Button'
 import FormControlLabel from '@mui/material/FormControlLabel'
-import FormGroup from '@mui/material/FormGroup'
-import FormHelperText from '@mui/material/FormHelperText'
 
-const signInSchema = object({
-  name: string()
-    .nonempty('Name is required')
-    .max(32, 'Name must be less than 100 characters'),
+const AuthLogInFormSchema = object({
   email: string().nonempty('Email is required').email('Email is invalid'),
   password: string()
     .nonempty('Password is required')
     .min(8, 'Password must be more than 8 characters')
     .max(32, 'Password must be less than 32 characters'),
-  passwordConfirm: string().nonempty('Please confirm your password'),
-  terms: literal(true, {
-    invalid_type_error: 'Accept Terms is required',
-  }),
-}).refine((data) => data.password === data.passwordConfirm, {
-  path: ['passwordConfirm'],
-  message: 'Passwords do not match',
 })
 
-type SignInInput = TypeOf<typeof signInSchema>
+type AuthLogInFormInput = TypeOf<typeof AuthLogInFormSchema>
 
-const SignInForm = () => {
+const AuthLogInForm = () => {
   const [loading, setLoading] = useState(false)
 
   const {
@@ -42,8 +30,8 @@ const SignInForm = () => {
     formState: { errors, isSubmitSuccessful },
     reset,
     handleSubmit,
-  } = useForm<SignInInput>({
-    resolver: zodResolver(signInSchema),
+  } = useForm<AuthLogInFormInput>({
+    resolver: zodResolver(AuthLogInFormSchema),
   })
 
   useEffect(() => {
@@ -53,15 +41,15 @@ const SignInForm = () => {
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [isSubmitSuccessful])
 
-  const onSubmitHandler: SubmitHandler<SignInInput> = (values) => {
+  const onSubmitHandler: SubmitHandler<AuthLogInFormInput> = (values) => {
     console.log(values)
   }
-  console.log(errors)
+  // console.log(errors)
 
   return (
     <Fragment>
       <Typography component="h1" variant="h5">
-        Sign up
+        Sign in
       </Typography>
       <Box
         component="form"
@@ -72,23 +60,10 @@ const SignInForm = () => {
         <TextField
           margin="normal"
           fullWidth
-          id="name"
-          label="Full Name"
-          placeholder="Full Name"
-          autoComplete="name"
-          autoFocus
-          required
-          {...register('name')}
-          error={!!errors['name']}
-          helperText={errors['name'] ? errors['name'].message : ''}
-        />
-        <TextField
-          margin="normal"
-          fullWidth
           id="email"
           label="Email Address"
-          placeholder="Email Address"
           autoComplete="email"
+          autoFocus
           required
           {...register('email')}
           error={!!errors['email']}
@@ -101,40 +76,12 @@ const SignInForm = () => {
           type="password"
           id="password"
           placeholder="Password"
-          autoComplete="new-password"
+          autoComplete="current-password"
           required
           error={!!errors['password']}
           helperText={errors['password'] ? errors['password'].message : ''}
           {...register('password')}
         />
-        <TextField
-          margin="normal"
-          fullWidth
-          label="passwordConfirm"
-          type="password"
-          id="passwordConfirm"
-          placeholder="password Confirm"
-          autoComplete="new-password"
-          required
-          error={!!errors['passwordConfirm']}
-          helperText={errors['passwordConfirm'] ? errors['passwordConfirm'].message : ''}
-          {...register('passwordConfirm')}
-        />
-
-        <FormGroup>
-          <FormControlLabel
-            control={<Checkbox required />}
-            {...register('terms')}
-            label={
-              <Typography color={errors['terms'] ? 'error' : 'inherit'}>
-                Accept Terms and Conditions
-              </Typography>
-            }
-          />
-          <FormHelperText error={!!errors['terms']}>
-            {errors['terms'] ? errors['terms'].message : ''}
-          </FormHelperText>
-        </FormGroup>
 
         <LoadingButton
           variant="contained"
@@ -143,14 +90,17 @@ const SignInForm = () => {
           loading={loading}
           sx={{ mt: 3, mb: 2 }}
         >
-          SIGN UP
+          Sign In
         </LoadingButton>
         <Grid container>
           <Grid item xs>
+            {/* <Link href="#" variant="body2">
+              Forgot password?
+            </Link> */}
           </Grid>
           <Grid item>
             <Link href="#" variant="body2">
-              {"Already have an account? Sign in"}
+              {"Don't have an account? Sign Up"}
             </Link>
           </Grid>
         </Grid>
@@ -159,4 +109,4 @@ const SignInForm = () => {
   )
 }
 
-export default SignInForm
+export default AuthLogInForm
