+++ packages/ui/Auth/SignIn/index.tsx
@@ -1,1 +1,1 @@
-export * from './SignInForm'
+export {default as Form} from './Form';
