+++ packages/ui/Base/Theme/Provider/index.tsx
@@ -1,0 +1,1 @@
+export {default as Provider} from './Provider';
