+++ packages/ui/Base/Theme/index.tsx
@@ -1,0 +1,1 @@
+export * from './Provider';
