--- packages/ui/Ingredients/AddForm/index.tsx
@@ -1,1 +1,0 @@
-export {default as AddForm} from './AddForm';
