+++ packages/ui/Ingredients/Show/Table/Row/index.tsx
@@ -1,0 +1,1 @@
+export {default as Row} from './Row';
