+++ packages/ui/Ingredients/Show/Table/index.tsx
@@ -1,0 +1,1 @@
+export {default as Table} from './Table';
