+++ packages/ui/Ingredients/Show/index.tsx
@@ -1,0 +1,1 @@
+export * from './Table';
