+++ packages/ui/Ingredients/index.tsx
@@ -1,0 +1,3 @@
+export * as Show from './Show';
+export * as Add from './Add';
+export * as model from './model';
