+++ packages/ui/Ingredients/model/UNITS/index.ts
@@ -1,0 +1,10 @@
+enum UNITS {
+  g = 'g',
+  Kg = 'Kg',
+  lts = 'lts',
+  ml = 'ml',
+  piece = 'piece',
+}
+
+export default UNITS
+
