+++ packages/ui/Ingredients/model/UNITS/interface/index.ts
@@ -1,0 +1,5 @@
+import IngredientUnit from '../index'
+
+type IngredientUnitlStrings = keyof typeof IngredientUnit
+
+export default IngredientUnitlStrings
