+++ packages/ui/Ingredients/model/index.ts
@@ -1,0 +1,5 @@
+// export {default as interface} from './interface';
+export {default as MockUp} from './MockUp';
+export {default as UNITS} from './UNITS';
+export * from './Utils';
+export type {default as TYPE} from './interface'
