+++ packages/ui/Ingredients/model/interface/index.ts
@@ -1,0 +1,9 @@
+import IngredientUnit from "../UNITS"
+
+export default interface Ingredient {
+  id: number
+  name: string
+  unit: IngredientUnit
+  price_per_unit: number
+  quantity: number
+}
