+++ packages/ui/index.tsx
@@ -1,3 +1,6 @@
 // import * as React from "react";
 
-export * from "./Button";
+export * as Ingredients from "./Ingredients";
+export * as Auth from "./Auth";
+// export * as Layouts from "./Layouts";
+// export * as Widgets from "./Widgets";
